code/requirements.txt
torch==2.2.1
torchvision==0.17.1
Pillow==10.2.0
numpy==1.23.5
pandas==2.0.3
scikit-learn==1.3.2
opencv-python==4.9.0.80